+++ JDesignPatterns/src/dp/structural/Decorator/CarroLuxo.java
@@ -3,21 +3,21 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dp.structural.decorator;
+package dp.structural.Decorator;
 
 /**
  *
  * @author Mateus Araújo
  */
-public class CarroSport extends CarroDecorator {
+public class CarroLuxo extends CarroDecorator {
 
-    public CarroSport(ICarro c) {
+    public CarroLuxo(ICarro c) {
         super(c);
     }
 
     @Override
     public void montar() {
         super.montar();
-        System.out.print(" Adicionando características ao carro sport.");
+        System.out.print(" Adicionando características ao carro luxo.");
     }
 }
